Tests/LibWeb/Text/expected/DOM/Element-documentElement-scrollWidthHeight.txt

document.documentElement.scrollWidth: 1242
document.documentElement.scrollHeight: 1481
